added welcome page. AIME test page does not show problems until Reset Problems DB is clicked.
